+++ state.go
@@ -46,10 +46,9 @@ type world struct {
 	hash   uint64
 }
 
-func createState[T Numeric | bool | string](s State[T]) State[T] {
-	s.hash = s.Hash()
-
-	return s
+// Check compares world and states2 by their hash.
+func (world world) Check(world2 world) bool {
+	return world.hash == world2.hash
 }
 
 func (state State[T]) GetKey() StateKey {
@@ -129,109 +128,12 @@ func (state State[T]) Hash() uint64 {
 	return hash
 }
 
-// Distance calculates the distance between the current state value and a condition target value
-// Returns 0 if the condition is already satisfied, otherwise returns the numeric distance
-func (state State[T]) Distance(condition ConditionInterface) float32 {
-	// Check if the condition key matches
-	if state.Key != condition.GetKey() {
-		return 0
-	}
+// updateHashIncremental updates a hash by removing old state and adding new state
+func updateHashIncremental(currentHash uint64, oldStateHash, newStateHash uint64) uint64 {
+	currentHash ^= oldStateHash // Remove old
+	currentHash ^= newStateHash // Add new
 
-	// Handle different condition types
-	switch cond := condition.(type) {
-	case *Condition[int8]:
-		if v, ok := any(state.Value).(int8); ok {
-			return calculateNumericDistance(float64(v), float64(cond.Value), cond.Operator)
-		}
-	case *Condition[int]:
-		if v, ok := any(state.Value).(int); ok {
-			return calculateNumericDistance(float64(v), float64(cond.Value), cond.Operator)
-		}
-	case *Condition[uint8]:
-		if v, ok := any(state.Value).(uint8); ok {
-			return calculateNumericDistance(float64(v), float64(cond.Value), cond.Operator)
-		}
-	case *Condition[uint64]:
-		if v, ok := any(state.Value).(uint64); ok {
-			return calculateNumericDistance(float64(v), float64(cond.Value), cond.Operator)
-		}
-	case *Condition[float64]:
-		if v, ok := any(state.Value).(float64); ok {
-			return calculateNumericDistance(v, cond.Value, cond.Operator)
-		}
-	case *ConditionBool:
-		if v, ok := any(state.Value).(bool); ok {
-			if cond.Operator == EQUAL {
-				if v == cond.Value {
-					return 0
-				}
-				return 1
-			} else if cond.Operator == NOT_EQUAL {
-				if v != cond.Value {
-					return 0
-				}
-				return 1
-			}
-		}
-	case *ConditionString:
-		if v, ok := any(state.Value).(string); ok {
-			if cond.Operator == EQUAL {
-				if v == cond.Value {
-					return 0
-				}
-				return 1
-			} else if cond.Operator == NOT_EQUAL {
-				if v != cond.Value {
-					return 0
-				}
-				return 1
-			}
-		}
-	}
-
-	return 0
-}
-
-// calculateNumericDistance computes the distance for numeric conditions based on operator
-func calculateNumericDistance(current, target float64, op operator) float32 {
-	switch op {
-	case EQUAL:
-		if current < target {
-			return float32(target - current)
-		}
-		return float32(current - target)
-	case NOT_EQUAL:
-		if current == target {
-			return 1.0
-		}
-		return 0.0
-	case UPPER_OR_EQUAL:
-		if current < target {
-			return float32(target - current)
-		}
-		return 0.0
-	case UPPER:
-		if current <= target {
-			return float32(target - current + 1)
-		}
-		return 0.0
-	case LOWER_OR_EQUAL:
-		if current > target {
-			return float32(current - target)
-		}
-		return 0.0
-	case LOWER:
-		if current >= target {
-			return float32(current - target + 1)
-		}
-		return 0.0
-	}
-	return 0.0
-}
-
-// Check compares world and states2 by their hash.
-func (world world) Check(world2 world) bool {
-	return world.hash == world2.hash
+	return currentHash
 }
 
 func (statesData states) GetIndex(stateKey StateKey) int {
@@ -242,14 +144,6 @@ func (statesData states) GetIndex(stateKey StateKey) int {
 	}
 
 	return -1
-}
-
-// updateHashIncremental updates a hash by removing old state and adding new state
-func updateHashIncremental(currentHash uint64, oldStateHash, newStateHash uint64) uint64 {
-	currentHash ^= oldStateHash // Remove old
-	currentHash ^= newStateHash // Add new
-
-	return currentHash
 }
 
 type Sensor any
